backend/requirements.txt
django
djangorestframework
django-cors-headers
apscheduler
requests
elasticsearch
